Updates / fixes
* Printed some information for the user to display what is going on
* Separated the code dealing with adding the ssh key
* Updated the way the git repositories are updated: I get the name of the
remote branch the current branch is tracking (if it exists), I then check
if the local branch is behind the remote ; if so I start a git pull - and I
warn the user so that (s)he can fix the conflicts if there are any.
If there is no local branch that matches the name of the branch we are
looking for I check the remote branches to see if there is one that
contains said name. If so I create a local branch that is set to track
the remote selected and I pull.

There are still, at least, the following "problems":
* if several remote names match I don't know what happens (haha)
* I don't ask the user if (s)he is fine with pulling from the remote: in my
case I assume that what is on the remote is tested and does not break
anything (in short it's safe to pull)
* when I create the local branch I assume the name of the remote follows
the pattern "origin/branch_name" (that is a BIG ASSUMPTION)